+++ Core/Sources/ChatPlugin/AITerminalChatPlugin.swift
