+++ Core/Sources/ChatPlugin/CallAIFunction.swift
