+++ Core/Sources/ChatPlugin/TerminalChatPlugin.swift
